feature: Replaces unknown type error with auto-reference
When a type is not found in TypeProvider, a TypeReference is automatically created instead of throwing an error. This reference will be checked for existence during the GraphQL package's schema resolution. This allows for Schema builders to reference types that have not yet been defined, and does not require explicit TypeReferences to do so.